pyproject.toml
[tool.poetry]
name = "profiles"
version = "0.1.0"
description = ""
authors = ["Eugene Morozov <jmv@emorozov.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
ipython = "^8.18.0"
django-extensions = "^3.2.3"
django-filter = "^23.4"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
